Fix external parser list when new parsers not present in uctags are defined
In addition to listing external parsers inside the BuiltInParsers
list, these parsers also have to be declared inside parse_p.h,
otherwise for any parser which isn't part of uctags the compilation
fails.

To do this, we have to separate the macros to

EXTERNAL_PARSER_LIST_FILE - the file in which the parsers are declared
EXTERNAL_PARSER_LIST - macro inside EXTERNAL_PARSER_LIST_FILE which
lists the external parsers

The EXTERNAL_PARSER_LIST macro is then used both to declare the
parsers inside parse_p.h and adding parsers to the BuiltInParsers array.